refacto parsing, pow is f32

+++ src/compute.rs
@@ -38,14 +38,14 @@ fn get_simplified_operation(operation: Vec<Operation>) -> Vec<Operation> {
 pub fn resolve(operation: &Vec<Operation>) -> String {
     let higher_polynomial = operation
         .iter()
-        .fold(0, |a, b| if a > b.pow { a } else { b.pow });
+        .fold(0.0, |a, b| if a > b.pow { a } else { b.pow });
     println!("Polynomial degree: {}", higher_polynomial);
-    if higher_polynomial > 2 {
+    if higher_polynomial > 2.0 {
         return "The polynomial degree is stricly greater than 2, I can't solve.".to_owned();
     }
-    let pow2 = get_operation_value_from_pow(&operation, 2);
-    let pow1 = get_operation_value_from_pow(&operation, 1);
-    let constant = get_operation_value_from_pow(&operation, 0);
+    let pow2 = get_operation_value_from_pow(&operation, 2.0);
+    let pow1 = get_operation_value_from_pow(&operation, 1.0);
+    let constant = get_operation_value_from_pow(&operation, 0.0);
     if pow2 == 0.0 {
         linear_operation(pow1, constant)
     } else {
@@ -106,7 +106,7 @@ pub fn quadratic_operation(a: f32, b: f32, c: f32) -> String {
     }
 }
 
-fn get_operation_value_from_pow(operation: &Vec<Operation>, pow: i16) -> f32 {
+fn get_operation_value_from_pow(operation: &Vec<Operation>, pow: f32) -> f32 {
     match operation.iter().find(|v| v.pow == pow) {
         Some(v) => v.value,
         None => 0.0,
